server/requirements.txt
alembic
bcrypt
cffi
click
Flask
Flask-Cors
Flask-SQLAlchemy
Flask-Testing
gunicorn
itsdangerous
Jinja2
Mako
MarkupSafe
psycopg2
pycparser
PyJWT
python-dateutil
python-editor
six
SQLAlchemy
Werkzeug
